Finish basic refactoring of math modules
