+++ apsg/__main__.py
@@ -3,33 +3,68 @@
 
 
 """
-Runs the interactive shell.
+Run the package as application.
 """
 
 
 import code
 import pkg_resources
-
+from argparse import ArgumentParser
 try:
     import readline  # NOQA
 except ImportError:
     pass
 
-from pylab import *  # NOQA
-from apsg import *  # NOQA
+from pylab import *     # NOQA
+
+# For `help(apsg)` working in console.
+import apsg
+
+from apsg import *      # NOQA
+from apsg.math import * # NOQA
 
 
-def main():
-    banner = "+----------------------------------------------------------+\n"
-    banner += "    APSG toolbox "
-    banner += pkg_resources.require("apsg")[0].version
-    banner += " - http://ondrolexa.github.io/apsg\n"
-    banner += "+----------------------------------------------------------+"
+def run_console_mode():
+    """
+    Run the interactive console mode.
+    """
+    version = pkg_resources.require("apsg")[0].version
+
+    banner =  "+------------------------------------------------------------------------------+\n"
+    banner += "|                                  APSG " + version + (34 * " ") +            "|\n"
+    banner += "|                                                                              |\n"
+    banner += "|                       http://ondrolexa.github.io/apsg                        |\n"
+    banner += "+------------------------------------------------------------------------------+"
+
     vars = globals().copy()
     vars.update(locals())
+
     shell = code.InteractiveConsole(vars)
     shell.interact(banner=banner)
 
+
+def main(args=None):
+    """
+    Main entry point for your project.
+
+    Arguments:
+        args : list
+            A of arguments as if they were input in the command line. Leave it
+            None to use sys.argv.
+    """
+    parser = ArgumentParser(description="APSG toolbox")
+
+    parser.add_argument("-i", "--interactive", action='store_true', default=False, help=run_console_mode.__doc__)
+    # todo parser.add_argument("-V", "--version", action='version')
+
+    parser.set_defaults(func=lambda: parser.print_usage())
+
+    result = parser.parse_args(args)
+
+    if result.interactive:
+        run_console_mode()
+    else:
+        result.func()
 
 if __name__ == "__main__":
     main()
