+++ apsg/math/__init__.py
@@ -6,5 +6,6 @@ Matrix, Vector and Tensor algebra in low dimension.
 """
 
 
-from apsg.math.matrix import Matrix2, Matrix3
-from apsg.math.vector import Vector2, Vector3
+from apsg.math.tensor import Tensor
+from apsg.math.matrix import Matrix2, Matrix3, Matrix4
+from apsg.math.vector import Vector2, Vector3, Vector4
