+++ apsg/setting.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+
+
+__all__ = ("settings",)
+
+
+
+settings = dict(notation="dd",          # Default notation for Fol dd or rhr
+                vec2dd=False,           # Show Vec3 as plunge direction and plunge
+                precision=1e-12,        # Numerical precision for comparism
+                figsize=(8, 6))         # Default figure size
+
